requirements.txt
numpy>=1.24
Pillow>=10.0
PyYAML>=6.0
tqdm>=4.66
torch>=2.2
torchvision>=0.17
mlflow>=2.11
segmentation-models-pytorch>=0.3.3
timm>=0.9.2
